Quick and dirty refactoring of ClusterEventPeriodical
Registering at the event bus from within a periodic job causes all sorts
 of funny behaviour when launching inputs
